feat: Add console surprise
