+++ .github/workflows/container-images-cd.yml
@@ -33,8 +33,10 @@ jobs:
               with:
                   fetch-depth: 2
 
-            - name: Update git SHA
-              run: echo "GIT_SHA = '${GITHUB_SHA}'" > posthog/gitsha.py
+            - name: Override git.py
+              run: >
+                  echo "def get_git_commit(): return '${GITHUB_SHA}'" > posthog/git.py
+                  echo "def get_git_branch(): return '${GITHUB_REF_NAME}'" >> posthog/git.py
 
             - name: Set up Docker Buildx
               uses: docker/setup-buildx-action@v2
