--- posthog/gitsha.py
@@ -1,1 +1,0 @@
-GIT_SHA = "Undefined"
